+++ TP/src/StaticMesh.h
@@ -29,7 +29,7 @@ class StaticMesh : NonCopyable {
         StaticMesh(const MeshData& data);
 
         void draw() const;
-        void draw(size_t count) const;
+        void draw(int count) const;
 
         BoundingSphere boundingSphere() const;
